Setup for handling exceptions based on the US version
