+++ test/lexer/tokens-laxed.json
@@ -1,0 +1,4 @@
+{
+  "email-templates/session-cancelled-en_US/html.ejs": [ { "type": "evaluate", "value": "eventDate" } ],
+  "email-templates/session-cancelled-en_US/text.ejs": [ { "type": "evaluate", "value": "eventDate" } ]
+}
